\name{descriptor}
\alias{descriptor}
\docType{data}
\title{
Chemical Descriptors of All 101 Organic Compounds
}
\description{
A 101*2279 dataframe contains all the descriptors calculated from DRAGON for 101 compounds.
}
\usage{data(descriptor)}
\examples{
data(descriptor)
dim(descriptor)
descriptor[1:5,1:5]
}
