\name{MLRMPA-package}
\alias{MLRMPA-package}
\alias{MLRMPA}
\docType{package}
\title{
Multiple Linear Regression Model Population Analysis
}
\description{
This package allows to perform Multiple Linear Regression Model Population Analysis. It contains
 quickly deleting non-information and random selecting descriptors from the clustered descriptors and 
 generating 2D-QSAR models, based on Monte Carlo Sampling technique and stepwise MLR.
 As a result a pool of models can be developed for modeling and prediction. Meanwhile, the 
 predictive ability and robustness of the developed models will be illustrated using the following
 validation techniques: leave-one-out (LOO) cross-validation procedures, leave-group-out (LGO) 
 cross-validation procedures, squared correlation coefficient of test set and validation through 
 an external test set.
}
\details{
\tabular{ll}{
Package: \tab MLRMPA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-21\cr
License: \tab GPL-2\cr
}
}
\author{
Meihong Xie, Xiaoyun Zhang 

Maintainer: to complain to <xyzhang@lzu.edu.cn> or <xiemeih@163.com>
}
\references{
H. D. Li, Y. Z. Liang, Q. S. Xu, D. S. Cao, \emph{J. Chemomet.} \bold{2010}, \emph{24}, 418-423.
}

\keyword{ MLRMPA }
\seealso{
  \code{\link{ClustOfVar}} 
}
