\name{plotcor}
\alias{plotcor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Histogram Plot of Correlation Distribution 
}
\description{
Create a histogram plot with the correlation between descriptors and the activity.
}
\usage{
plotcor(descriptor.tr, activity.tr, maintitle, xlab, breaks, xlim)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{descriptor.tr}{
  a dataframe contained the descriptors of training set
}
  \item{activity.tr}{
  a dataframe contained the experimental data of training set
}
  \item{maintitle}{
  title of the plot
}
  \item{xlab}{
  a title for the x axis
}
  \item{breaks}{
  a single number giving the number of cells for the histogram
}
  \item{xlim}{
  the x limits of the plot
}
}

\keyword{documentation}
\keyword{utilities}

