\name{plotvar}
\alias{plotvar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Histogram Plot of Variance Distribution
}
\description{
Create a histogram plot with the varinace and activity.
}
\usage{
plotvar(descriptor.tr, activity.tr, maintitle, xlab, breaks)
}
\arguments{
  \item{descriptor.tr}{
  a dataframe contained the descriptors of training set
}
  \item{activity.tr}{
  a dataframe contained the experimental data of training set
}
  \item{maintitle}{
  title of the plot
}
  \item{xlab}{
  a title for the x axis
}
  \item{breaks}{
  a single number giving the number of cells for the histogram
}
}

\keyword{ documentation }
\keyword{ utilities }
